refactor: Перенос логики отрисовки каталога в отдельный контроллер
